services/notification-service/requirements.txt
pika
pymongo
requests
loguru
pydantic
pydantic-settings
python-dotenv
